+++ modules/carto/src/index.ts
@@ -1,5 +1,5 @@
 import {default as H3TileLayer} from './layers/h3-tile-layer';
-import {default as HeatmapTileLayer} from './layers/quadbin-heatmap-tile-layer';
+import {default as HeatmapTileLayer} from './layers/heatmap-tile-layer';
 import {default as _PointLabelLayer} from './layers/point-label-layer';
 import {default as QuadbinTileLayer} from './layers/quadbin-tile-layer';
 import {default as RasterTileLayer} from './layers/raster-tile-layer';
